structure adjustments (#3532)

+++ native/cocos/renderer/gfx-gles2/GLES2Device.h
@@ -86,7 +86,7 @@ public:
         });
     }
 
-    inline uint getThreadID() const { return _threadID; }
+    inline size_t getThreadID() const { return _threadID; }
 
 protected:
     static GLES2Device *instance;
@@ -133,7 +133,7 @@ protected:
 
     StringArray _extensions;
 
-    uint _threadID = 0U;
+    size_t _threadID = 0U;
 };
 
 } // namespace gfx
